IR 28: Un bot mas cortes Fixing some issues with arguments overlapping HTTP header names.
Changing the way exceptions are being managed.

Now, each command in the Commands class does handle it's own excpetions so they can notify the UserInteraction class about it.

Cambios en los threads, faltan las exceptions.

Working on the exceptions handling for the UserInteraction usage.

Handling exceptions inside the Commands class to notify the UserInteraction instance running that an error has ocurred so it notifies that to the user at
Slack.

Finishing the UserInteraction class stuff.