IR-3 - Recomendaciones
Potentially finished this recommendation command.

Added the usage of --team in the compare commands

Now you can specify a teamname and the bot will get the usernames from the
team, if the team does not exist or there were found no contestants an error
message will be sent to slack.

Fixing a potential error in the recommendations command

The compare command now shows the usernames when passed a team.